move otf and lineup logic into states to separate from drive
