+++ src/main/java/frc/robot/BuildConstants.java
@@ -5,12 +5,12 @@ public final class BuildConstants {
   public static final String MAVEN_GROUP = "";
   public static final String MAVEN_NAME = "2025-Robot-Code";
   public static final String VERSION = "unspecified";
-  public static final int GIT_REVISION = 55;
-  public static final String GIT_SHA = "09aafdb8016cc629e0daebca2bb28702dcb62ea7";
-  public static final String GIT_DATE = "2025-02-03 18:32:50 EST";
-  public static final String GIT_BRANCH = "landing-zone-alignment";
-  public static final String BUILD_DATE = "2025-02-03 19:20:28 EST";
-  public static final long BUILD_UNIX_TIME = 1738628428795L;
+  public static final int GIT_REVISION = 18;
+  public static final String GIT_SHA = "68678ee8f872b701c88c825bca64cbfb12edde36";
+  public static final String GIT_DATE = "2025-02-03 19:26:05 EST";
+  public static final String GIT_BRANCH = "drive-state";
+  public static final String BUILD_DATE = "2025-02-04 12:34:01 EST";
+  public static final long BUILD_UNIX_TIME = 1738690441467L;
   public static final int DIRTY = 1;
 
   private BuildConstants() {}
