+++ src/main/java/frc/robot/subsystems/drive/Drive.java
@@ -9,17 +9,14 @@ import com.pathplanner.lib.config.ModuleConfig;
 import com.pathplanner.lib.config.PIDConstants;
 import com.pathplanner.lib.config.RobotConfig;
 import com.pathplanner.lib.controllers.PPHolonomicDriveController;
-import com.pathplanner.lib.path.PathConstraints;
 import com.pathplanner.lib.pathfinding.Pathfinding;
 import com.pathplanner.lib.util.PathPlannerLogging;
-import coppercore.parameter_tools.LoggedTunableNumber;
 import coppercore.vision.VisionLocalizer.DistanceToTag;
 import coppercore.wpilib_interface.DriveTemplate;
 import edu.wpi.first.hal.FRCNetComm.tInstances;
 import edu.wpi.first.hal.FRCNetComm.tResourceType;
 import edu.wpi.first.hal.HAL;
 import edu.wpi.first.math.Matrix;
-import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.math.estimator.SwerveDrivePoseEstimator;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
@@ -32,8 +29,6 @@ import edu.wpi.first.math.kinematics.SwerveModuleState;
 import edu.wpi.first.math.numbers.N1;
 import edu.wpi.first.math.numbers.N3;
 import edu.wpi.first.math.system.plant.DCMotor;
-import edu.wpi.first.math.trajectory.TrapezoidProfile;
-import edu.wpi.first.math.trajectory.TrapezoidProfile.Constraints;
 import edu.wpi.first.networktables.DoubleSubscriber;
 import edu.wpi.first.networktables.NetworkTable;
 import edu.wpi.first.networktables.NetworkTableInstance;
@@ -45,8 +40,8 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
 import frc.robot.Constants;
 import frc.robot.Constants.Mode;
-import frc.robot.TestModeManager;
 import frc.robot.constants.JsonConstants;
+import frc.robot.subsystems.drive.states.OTFState;
 import frc.robot.util.LocalADStarAK;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
@@ -154,71 +149,6 @@ public class Drive implements DriveTemplate {
 
   private DesiredLocation desiredLocation = DesiredLocation.Reef0;
 
-  private boolean isOTF = false;
-
-  private boolean isLiningUp = false;
-
-  private Command driveToPose = null;
-
-  private VisionAlignment alignmentSupplier;
-
-  // along track pid test mode
-  private LoggedTunableNumber alongTrackkP =
-      new LoggedTunableNumber(
-          "DriveLineupGains/AlongTrackkP", JsonConstants.drivetrainConstants.driveAlongTrackkP);
-  private LoggedTunableNumber alongTrackkI =
-      new LoggedTunableNumber(
-          "DriveLineupGains/AlongTrackkI", JsonConstants.drivetrainConstants.driveAlongTrackkI);
-  private LoggedTunableNumber alongTrackkD =
-      new LoggedTunableNumber(
-          "DriveLineupGains/AlongTrackkD", JsonConstants.drivetrainConstants.driveAlongTrackkD);
-
-  // cross tack pid test mode
-  private LoggedTunableNumber crossTrackkP =
-      new LoggedTunableNumber(
-          "DriveLineupGains/CrossTrackkP", JsonConstants.drivetrainConstants.driveCrossTrackkP);
-  private LoggedTunableNumber crossTrackkI =
-      new LoggedTunableNumber(
-          "DriveLineupGains/CrossTrackkI", JsonConstants.drivetrainConstants.driveCrossTrackkI);
-  private LoggedTunableNumber crossTrackkD =
-      new LoggedTunableNumber(
-          "DriveLineupGains/CrossTrackkD", JsonConstants.drivetrainConstants.driveCrossTrackkD);
-
-  // rotation pid test mode
-  private LoggedTunableNumber rotationkP =
-      new LoggedTunableNumber(
-          "DriveLineupGains/rotationkP", JsonConstants.drivetrainConstants.driveRotationkP);
-  private LoggedTunableNumber rotationkI =
-      new LoggedTunableNumber(
-          "DriveLineupGains/rotationkI", JsonConstants.drivetrainConstants.driveRotationkI);
-  private LoggedTunableNumber rotationkD =
-      new LoggedTunableNumber(
-          "DriveLineupGains/rotationkD", JsonConstants.drivetrainConstants.driveRotationkD);
-
-  private PIDController driveAlongTrackLineupController =
-      new PIDController(
-          JsonConstants.drivetrainConstants.driveAlongTrackkP,
-          JsonConstants.drivetrainConstants.driveAlongTrackkI,
-          JsonConstants.drivetrainConstants.driveAlongTrackkD);
-  private PIDController driveCrossTrackLineupController =
-      new PIDController(
-          JsonConstants.drivetrainConstants.driveCrossTrackkP,
-          JsonConstants.drivetrainConstants.driveCrossTrackkI,
-          JsonConstants.drivetrainConstants.driveCrossTrackkD);
-
-  private Constraints driveAlongTrackProfileConstraints =
-      new Constraints(
-          JsonConstants.drivetrainConstants.driveAlongTrackVelocity,
-          JsonConstants.drivetrainConstants.driveAlongTrackVelocity);
-  private TrapezoidProfile driveAlongTrackProfile =
-      new TrapezoidProfile(driveAlongTrackProfileConstraints);
-
-  private PIDController rotationController =
-      new PIDController(
-          JsonConstants.drivetrainConstants.driveRotationkP,
-          JsonConstants.drivetrainConstants.driveRotationkI,
-          JsonConstants.drivetrainConstants.driveRotationkD);
-
   private NetworkTableInstance inst = NetworkTableInstance.getDefault();
   private NetworkTable table = inst.getTable("");
   private DoubleSubscriber reefLocationSelector = table.getDoubleTopic("reefTarget").subscribe(-1);
@@ -229,7 +159,6 @@ public class Drive implements DriveTemplate {
       ModuleIO frModuleIO,
       ModuleIO blModuleIO,
       ModuleIO brModuleIO) {
-    this.alignmentSupplier = null;
     this.gyroIO = gyroIO;
     modules[0] = new Module(flModuleIO, 0, DriveConfiguration.getInstance().FrontLeft);
     modules[1] = new Module(frModuleIO, 1, DriveConfiguration.getInstance().FrontRight);
@@ -277,8 +206,6 @@ public class Drive implements DriveTemplate {
                 (state) -> Logger.recordOutput("Drive/SysIdState", state.toString())),
             new SysIdRoutine.Mechanism(
                 (voltage) -> runCharacterization(voltage.in(Volts)), null, this));
-
-    rotationController.enableContinuousInput(-Math.PI / 2, Math.PI / 2);
   }
 
   @Override
@@ -304,24 +231,8 @@ public class Drive implements DriveTemplate {
       Logger.recordOutput("SwerveStates/SetpointsOptimized", new SwerveModuleState[] {});
     }
 
-    // OTF Command
-    if (driveToPose != null) {
-      Logger.recordOutput("Drive/OTF/OnTheFlyCommandStatus", this.driveToPose.isScheduled());
-
-      // cancel path following command once OTF cancelled (likely via trigger)
-      if (!isOTF) {
-        driveToPose.cancel();
-      }
-    }
-
-    if (isDriveCloseToFinalLineupPose() && isOTF) {
-      this.setOTF(false);
-      driveToPose.cancel();
-      this.setLiningUp(true);
-    }
-
-    if (isLiningUp) {
-      this.LineupWithReefLocation();
+    if (isDriveCloseToFinalLineupPose() && isDriveOTF()) {
+      // TODO: auto transition to LineupState?
     }
 
     // check for update from reef touchscreen
@@ -368,39 +279,6 @@ public class Drive implements DriveTemplate {
     gyroDisconnectedAlert.set(!gyroInputs.connected && Constants.currentMode != Mode.SIM);
   }
 
-  public void testPeriodic() {
-    switch (TestModeManager.getTestMode()) {
-      case DriveLineupTuning:
-        LoggedTunableNumber.ifChanged(
-            hashCode(),
-            (pid) -> {
-              this.setAlongTrackPID(pid[0], pid[1], pid[2]);
-            },
-            alongTrackkP,
-            alongTrackkI,
-            alongTrackkD);
-        LoggedTunableNumber.ifChanged(
-            hashCode(),
-            (pid) -> {
-              this.setCrossTrackPID(pid[0], pid[1], pid[2]);
-            },
-            crossTrackkP,
-            crossTrackkI,
-            crossTrackkD);
-        LoggedTunableNumber.ifChanged(
-            hashCode(),
-            (pid) -> {
-              this.setRotationLineupPID(pid[0], pid[1], pid[2]);
-            },
-            rotationkP,
-            rotationkI,
-            rotationkD);
-        break;
-      default:
-        break;
-    }
-  }
-
   /**
    * sets desired speeds of robot
    *
@@ -428,73 +306,14 @@ public class Drive implements DriveTemplate {
   }
 
   /**
-   * sets lineup along track pid gains
-   *
-   * @param kP proportional gain
-   * @param kI integral gain
-   * @param kD derivative gain
-   */
-  public void setAlongTrackPID(double kP, double kI, double kD) {
-    this.driveAlongTrackLineupController = new PIDController(kP, kI, kD);
-  }
-
-  /**
-   * sets lineup cross track pid gains
-   *
-   * @param kP proportional gain
-   * @param kI integral gain
-   * @param kD derivative gain
-   */
-  public void setCrossTrackPID(double kP, double kI, double kD) {
-    this.driveCrossTrackLineupController = new PIDController(kP, kI, kD);
-  }
-
-  /**
-   * sets lineup rotation pid gains
-   *
-   * @param kP proportional gain
-   * @param kI integral gain
-   * @param kD derivative gain
-   */
-  public void setRotationLineupPID(double kP, double kI, double kD) {
-    this.rotationController = new PIDController(kP, kI, kD);
-    this.rotationController.enableContinuousInput(-Math.PI / 2, Math.PI / 2);
-  }
-
-  /**
-   * sets the supplier for landing zone alignment help
-   *
-   * @param alignmentSupplier from vision.getDistanceErrorToTag; helps drive align to reef
-   */
-  public void setAlignmentSupplier(VisionAlignment alignmentSupplier) {
-    this.alignmentSupplier = alignmentSupplier;
-  }
-
-  /**
-   * sets isOTF of robot true will cause robot to create a path from current location to the set
-   * PathLocation
-   *
-   * @param isOTF boolean telling robot if it should create a OTF path
-   */
-  public void setOTF(boolean isOTF) {
-    this.isOTF = isOTF;
-
-    if (isOTF) {
-      this.driveToPose = this.getDriveToPoseCommand();
-      if (this.driveToPose != null) {
-        this.driveToPose.schedule();
-      }
-    }
-  }
-
-  /**
    * checks if drive is currently following an on the fly path
    *
    * @return state of OTF following
    */
   @AutoLogOutput(key = "Drive/OTF/isOTF")
   public boolean isDriveOTF() {
-    return isOTF;
+    // TODO: check for state
+    return false;
   }
 
   /**
@@ -506,20 +325,10 @@ public class Drive implements DriveTemplate {
     // relative to transforms first pose into distance from desired pose
     // then get distance between poses (if less than 0.1 meters we are good)
     return this.getPose()
-            .relativeTo(this.findOTFPoseFromDesiredLocation())
+            .relativeTo(OTFState.findOTFPoseFromDesiredLocation(this))
             .getTranslation()
             .getNorm()
         < JsonConstants.drivetrainConstants.otfPoseDistanceLimit;
-  }
-
-  /**
-   * sets isLiningUp of robot true will cause robot to gather distance from reef tag and drive
-   * towards it PathLocation
-   *
-   * @param isLiningUp boolean telling robot if it should create a OTF path
-   */
-  public void setLiningUp(boolean isLiningUp) {
-    this.isLiningUp = isLiningUp;
   }
 
   /**
@@ -529,7 +338,8 @@ public class Drive implements DriveTemplate {
    */
   @AutoLogOutput(key = "Drive/Lineup/isLiningUp")
   public boolean isDriveLiningUp() {
-    return isLiningUp;
+    // TODO: return state if lining up
+    return false;
   }
 
   /**
@@ -544,24 +354,6 @@ public class Drive implements DriveTemplate {
   }
 
   /**
-   * allows drive to be controlled by on the fly / landing zone alignment
-   *
-   * @param autoAlignment true allows drive to go into otf and alignment
-   */
-  public void setAutoAlignment(boolean autoAlignment) {
-    if (autoAlignment) {
-      if (isDriveCloseToFinalLineupPose() && isDesiredLocationReef()) {
-        this.setLiningUp(true);
-      } else {
-        this.setOTF(true);
-      }
-    } else {
-      setOTF(false);
-      setLiningUp(false);
-    }
-  }
-
-  /**
    * sets desired path location calling this and then setting OTF to true will cause robot to drive
    * path from current pose to the location
    *
@@ -569,6 +361,13 @@ public class Drive implements DriveTemplate {
    */
   public void setDesiredLocation(DesiredLocation location) {
     this.desiredLocation = location;
+  }
+
+  /**
+   * @return the desired location for otf and lineup
+   */
+  public DesiredLocation getDesiredLocation() {
+    return this.desiredLocation;
   }
 
   /**
@@ -588,7 +387,7 @@ public class Drive implements DriveTemplate {
       return;
     }
     if (locationArray[(int) desiredIndex] != desiredLocation) {
-      if (isOTF) {
+      if (isDriveOTF()) {
         this.updateDesiredLocation((int) desiredIndex);
       } else {
         this.setDesiredLocation((int) desiredIndex);
@@ -616,61 +415,9 @@ public class Drive implements DriveTemplate {
   public void updateDesiredLocation(DesiredLocation location) {
     this.setDesiredLocation(location);
 
-    if (isOTF) {
-      this.driveToPose.cancel();
-      this.driveToPose = this.getDriveToPoseCommand();
-      this.driveToPose.schedule();
+    if (isDriveOTF()) {
+      // TODO: deal with state (re run on entry?)
     }
-  }
-
-  /**
-   * finds a pose to pathfind to based on desiredLocation enum
-   *
-   * @return a pose representing the corresponding scoring location
-   */
-  public Pose2d findOTFPoseFromDesiredLocation() {
-    switch (this.desiredLocation) {
-        // NOTE: pairs of reef sides (ie 0 and 1) will have the same otf pose (approximately 0.5-1
-        // meter away from center of tag)
-      case Reef0:
-        return DriverStation.getAlliance().orElse(Alliance.Blue) == Alliance.Red
-            ? JsonConstants.redFieldLocations.reef0
-            : new Pose2d();
-      case Reef1:
-        return new Pose2d(Meters.of(14.350), Meters.of(4.0), new Rotation2d(Degrees.of(180)));
-      case CoralStationRight:
-        return new Pose2d(16.0, 6.6, new Rotation2d(0.0));
-        // return new Pose2d(1.2, 1, Rotation2d.fromRadians(1));
-      case CoralStationLeft:
-        return new Pose2d(1.2, 7.0, Rotation2d.fromRadians(-1));
-      default:
-        // no location set, so don't allow drive to run OTF
-        this.setOTF(false);
-        return null;
-    }
-  }
-
-  /**
-   * gets the path from current pose to the desired pose found from location
-   *
-   * @return command that drive can schedule to follow the path found
-   */
-  public Command getDriveToPoseCommand() {
-    Pose2d targetPose = findOTFPoseFromDesiredLocation();
-
-    if (targetPose == null) {
-      return null;
-    }
-
-    // Create the constraints to use while pathfinding
-    PathConstraints constraints =
-        new PathConstraints(
-            JsonConstants.drivetrainConstants.OTFMaxLinearVelocity,
-            JsonConstants.drivetrainConstants.OTFMaxLinearAccel,
-            JsonConstants.drivetrainConstants.OTFMaxAngularVelocity,
-            JsonConstants.drivetrainConstants.OTFMaxAngularAccel);
-
-    return AutoBuilder.pathfindToPose(targetPose, constraints, 0.0);
   }
 
   /**
@@ -680,178 +427,6 @@ public class Drive implements DriveTemplate {
    */
   public boolean isAllianceRed() {
     return DriverStation.getAlliance().orElse(Alliance.Blue).equals(Alliance.Red);
-  }
-
-  /**
-   * gets tag to use for final alignment with vision
-   *
-   * @return int representing tag id to use
-   */
-  public int getTagIdForReef() {
-    boolean allianceRed = this.isAllianceRed();
-    switch (desiredLocation) {
-      case Reef0:
-        return 7;
-      case Reef1:
-        return allianceRed ? 10 : 21;
-      case Reef2:
-      case Reef3:
-        return allianceRed ? 9 : 22;
-      case Reef4:
-      case Reef5:
-        return allianceRed ? 8 : 17;
-      case Reef6:
-      case Reef7:
-        return allianceRed ? 7 : 18;
-      case Reef8:
-      case Reef9:
-        return allianceRed ? 6 : 19;
-      case Reef10:
-      case Reef11:
-        return allianceRed ? 11 : 20;
-      default:
-        return -1;
-    }
-  }
-
-  /**
-   * gets cross track offset for lineup
-   *
-   * @param cameraIndex camera to check offset
-   * @return offset for camera
-   */
-  public Double getCrossTrackOffset(int cameraIndex) {
-    if (cameraIndex == 0) {
-      return JsonConstants.drivetrainConstants.driveCrossTrackFrontRightOffset;
-    }
-    return 0.0; // front left offset (not added)
-  }
-
-  /**
-   * gets camera index for vision single tag lineup
-   *
-   * @return 0 for Front Left camera; 1 for Front Right camera
-   */
-  public int getCameraIndexForLineup() {
-    switch (desiredLocation) {
-        // Right Side of reef side (align to left camera)
-      case Reef0:
-      case Reef2:
-      case Reef4:
-      case Reef6:
-      case Reef8:
-      case Reef10:
-        return JsonConstants.visionConstants.FrontLeftCameraIndex;
-        // Left side of reef side (align to right camera)
-      case Reef1:
-      case Reef3:
-      case Reef5:
-      case Reef7:
-      case Reef9:
-      case Reef11:
-        return JsonConstants.visionConstants.FrontRightCameraIndex;
-      default:
-        return -1;
-    }
-  }
-
-  /**
-   * gets rotation for each side of hexagonal reef for lineup
-   *
-   * @return Rotation2d representing desired rotation for lineup
-   */
-  public Rotation2d getRotationForReefSide() {
-    switch (desiredLocation) {
-      case Reef0:
-        return new Rotation2d();
-      case Reef1:
-        return new Rotation2d();
-      case Reef2:
-        return new Rotation2d();
-      case Reef3:
-        return new Rotation2d();
-      case Reef4:
-        return new Rotation2d();
-      case Reef5:
-        return new Rotation2d();
-      case Reef6:
-        return new Rotation2d();
-      case Reef7:
-        return new Rotation2d();
-      case Reef8:
-        return new Rotation2d();
-      case Reef9:
-        return new Rotation2d();
-      case Reef10:
-        return new Rotation2d();
-      case Reef11:
-        return new Rotation2d();
-      default:
-        return new Rotation2d();
-    }
-  }
-
-  private DistanceToTag latestObservation;
-  private int observationAge;
-
-  /** take over goal speeds to align to reef exactly */
-  public void LineupWithReefLocation() {
-    int tagId = this.getTagIdForReef();
-    int cameraIndex = this.getCameraIndexForLineup();
-
-    if (tagId == -1 || cameraIndex == -1 || alignmentSupplier == null) {
-      // cancel lineup or whatever
-      this.setLiningUp(false);
-      return;
-    }
-
-    System.out.println(tagId);
-
-    DistanceToTag observation =
-        alignmentSupplier.get(
-            tagId,
-            cameraIndex,
-            this.getCrossTrackOffset(cameraIndex),
-            JsonConstants.drivetrainConstants.driveAlongTrackOffset);
-
-    if (!observation.isValid()) {
-      if (latestObservation != null && observationAge < 5) {
-        observation = latestObservation;
-        observationAge++;
-      } else {
-        return;
-      }
-    } else {
-      latestObservation = observation;
-      observationAge = 0;
-    }
-
-    Logger.recordOutput("Drive/Lineup/AlongTrackDistance", observation.alongTrackDistance());
-    Logger.recordOutput("Drive/Lineup/CrossTrackDistance", observation.crossTrackDistance());
-    Logger.recordOutput("Drive/Lineup/IsObservationValid", observation.isValid());
-
-    // give to PID Controllers and setGoalSpeeds (robotCentric)
-    double vx =
-        JsonConstants.drivetrainConstants.driveAlongTrackMultiplier
-            * driveAlongTrackLineupController.calculate(observation.alongTrackDistance());
-    // ChassisSpeeds speeds = getChassisSpeeds();
-    // Pose2d pose = getPose();
-    // double velocityToTarget = (pose.getX() * speeds.vxMetersPerSecond + pose.getY() *
-    // speeds.vyMetersPerSecond) / Math.sqrt(pose.getX() * pose.getX() + pose.getY() * pose.getY());
-    // double velocityToTarget = -getChassisSpeeds().vxMetersPerSecond;
-    // Logger.recordOutput("Drive/Lineup/velocityToTarget", velocityToTarget);
-    // double vx =
-    //     -driveAlongTrackProfile.calculate(
-    //             0.02,
-    //             new State(observation.alongTrackDistance(), velocityToTarget),
-    //             new State(0, 0))
-    //         .velocity;
-    double vy = driveCrossTrackLineupController.calculate(observation.crossTrackDistance());
-    double omega =
-        rotationController.calculate(
-            this.getRotation().getRadians(), this.getRotationForReefSide().getRadians());
-
-    this.setGoalSpeeds(new ChassisSpeeds(vx, vy, omega), false);
   }
 
   /** Runs the drive at the desired speeds set in (@Link setGoalSpeeds) */
